add both wandb and tensorboard support

+++ model/helper/common.py
@@ -34,7 +34,7 @@ def clusters_voxelization(clusters_idx, clusters_offset, feats, coords, scale, s
     clusters_coords += offset
     assert clusters_coords.shape.numel() == ((clusters_coords >= 0) * (clusters_coords < spatial_shape)).sum()
 
-    clusters_coords = clusters_coords.long().cpu()
+    clusters_coords = clusters_coords.int().cpu()
 
     clusters_voxel_coords, clusters_p2v_map, clusters_v2p_map = common_ops.voxelization_idx(clusters_coords,
                                                                                             clusters_idx[:, 0].to(torch.int16),
